+++ crates/entrypoint/src/cli/command/azure_devops/mod.rs
@@ -11,6 +11,7 @@ pub mod service_endpoint;
 pub mod team;
 pub mod test;
 pub mod work_item_query;
+pub mod user;
 
 use crate::cli::azure_devops::azure_devops_command::AzureDevOpsCommand;
 use clap::Args;