+++ crates/entrypoint/src/cli/command/azure_devops/user/mod.rs
@@ -1,0 +1,5 @@
+pub mod azure_devops_user_cli;
+pub mod update;
+
+pub use azure_devops_user_cli::*;
+pub use update::*;