+++ crates/entrypoint/src/cli/command/azure_devops/user/update/mod.rs
@@ -1,0 +1,5 @@
+pub mod azure_devops_user_update_cli;
+pub mod azure_devops_user_update_tui;
+
+pub use azure_devops_user_update_cli::*;
+pub use azure_devops_user_update_tui::*;